requirements.txt
flask
flask_sqlalchemy
flask_wtf
flask_bcrypt
flask_login
flask_socketio
email_validator
simple-websocket
python-dateutil
phonenumbers
gevent
gevent-websocket